Updates to scenarios

+++ backend/src/main/java/com/companyA/backend/InventoryManagementSystem/contoller/StockAlertController.java
@@ -17,20 +17,24 @@ public class StockAlertController {
 
     private final StockAlertService stockAlertService;
 
+    //Dependency Injection
     @Autowired
     public StockAlertController(StockAlertService stockAlertService) {
         this.stockAlertService = stockAlertService;
     }
 
-    //view all stock alerts
+    //View all stock alerts
     @GetMapping
     public ResponseEntity<List<StockAlert>> getAllStockAlerts() {
         return new ResponseEntity<List<StockAlert>>(stockAlertService.allStockAlerts(), HttpStatus.OK);
     }
 
-    //Remove a stock alert
+    //Remove a stock alert by stock alert Id
     @DeleteMapping("/deleteStockAlert/{stockAlertId}")
     public ResponseEntity<String> deleteStockAlert(@PathVariable String stockAlertId) {
+        if(stockAlertId == null || stockAlertId.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         return new ResponseEntity<String>(stockAlertService.deleteStockAlert(stockAlertId),HttpStatus.OK);
     }
